+++ react/__mocks__/vtex.responsive-values.ts
@@ -1,0 +1,5 @@
+export const useResponsiveValue = (font: {
+  desktop?: string
+  mobile?: string
+  tablet?: string
+}) => font.desktop || font.mobile || font.tablet || font